Add another version of Hash Table
Add another version of a Hash Table that is able to dynamically allocate
memory space within itself, however it is a global variable, rather than
one that can be easily recreated for use elsewhere and has strong type
restrictions with its Key and Value types (which is to be expected).

Ultimately this was just a test to see if it worked and further
evaluate if it is worth using going forward, or what can be done to make
it more reusable using it as a base to work from.

== References:
- [Hash from two arrays \- Rosetta Code]
  (https://rosettacode.org/wiki/Hash_from_two_arrays#C)