Add Robot move function
Add a Robot move function that will update either the `x` or `y` integer
values within the Robot struct, depending on which direction it is
facing; i.e.

- When facing North, `y` increase by 1
- When facing South, `y` increase by -1
- When facing East, `x` increase by 1
- When facing West, `x` increase by -1

Otherwise if not specified above, the `x` or `y` value does not change.

Also, if the value of either `x` or `y` goes over 4 or under 0, it will
be reset to the last known position, in order to adhere to the Toy Robot
specification.

== Notes:
- Remove unnecessary comments

- > Structs in C are passed by value, not by reference.
  This was not initially apparent until trying to reassign values within
  a passed in Struct, but not when copying values into a `char*` array

== References:
- [C library function \- atoi\(\)]
  (https://www.tutorialspoint.com/c_standard_library/c_function_atoi.htm)

- [c \- How to change values of a struct passed to a function \- Stack
  Overflow]
  (https://stackoverflow.com/questions/15538711/how-to-change-values-of-a-struct-passed-to-a-function)

- [C program to convert string to integer without using atoi function \| Programming Simplified]
  (https://www.programmingsimplified.com/c/source-code/c-program-convert-string-to-integer-without-using-atoi-function)