Add Robot place function
Add a place function within the Robot library, which accepts a pointer
to a Robot instance location and `char*` array that contains raw
coordinates on where to place the Robot; i.e. change its `x`, `y` &
`facing` direction; in the form of "0,0,NORTH".

This fulfils the Place functionality requirement of the Toy Robot
specification.

== References:
- [c \- Get a substring of a char\* \- Stack Overflow]
  (https://stackoverflow.com/questions/4214314/get-a-substring-of-a-char)