Add OS conditional to compile BSD specific library
Add an operating system conditional to only compile an extra library
that is only need when compiling on a macOS system, built on-top of BSD
OS, due to how it implements the `fmemopen` function used in tests. This
conditional relies on getting the OS name from a shell command and then
adding the additional library used in tests.

== References:
- [makefile \- How to detect OS in a make file? \- Stack Overflow]
  (https://stackoverflow.com/questions/9279765/how-to-detect-os-in-a-make-file)

- [GNU make: Conditional Example]
  (https://www.gnu.org/software/make/manual/html_node/Conditional-Example.html#Conditional-Example)